# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "EEducation API service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1234",
    "basePath": "/",
    "paths": {
...
        "/admin/profiles": {
            "delete": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Delete user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete user \\\\ ADMIN ONLY",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Change user activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Change user activity \\\\ ADMIN ONLY",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.ChangeUserStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
        "/profile/edit": {
            "put": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Fill user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Fill user profile",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.FillUserProfileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
        "/schedule": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get weekly schedule by class",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get weekly schedule by class",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request",
                        "name": "classname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Class not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
        "/teacher/lessons/set-grades": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get students by classname",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get students by classname",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request",
                        "name": "classname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Class not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
        "/teacher/schedule/set-schedule/import": {
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get weekly schedule by class",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get weekly schedule by class",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Scheudle excel file",
                        "name": "uploaded-schedule",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sheet",
                        "name": "sheet",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
        "/teacher/schedule/set-schedule/lessons": {
            "patch": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Set homework",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Set homework",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schedule.SetHomeworkReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Homework updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Lesson not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "schedule.SetHomeworkReq": {
            "type": "object",
            "properties": {
                "classname": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "homework": {
                    "type": "string"
                },
                "lesson_name": {
                    "type": "string"
                }
            }
        },
        "user.ChangeUserStatusReq": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "initiator_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "user.FillUserProfileReq": {
            "type": "object",
            "properties": {
                "classname": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "TokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}